refactor(naming): loop_a/b/c를 자기서술적 이름으로 리네임 (하위호환 shim 유지)
암호 같던 loop_a/loop_b/loop_c를 canonical env 접두사와 일치하는
설명적 이름으로 통일:
- tools/loop_a_hardstop.py   -> tools/hardstop_seller.py   (HARDSTOP_*)
- tools/loop_b_trend_exit.py -> tools/trend_exit_seller.py (TREND_EXIT_*)
- tools/loop_c_fill_chaser.py-> tools/fill_chaser.py       (FILL_CHASER_*)
- loop_publish.py            -> sell_broadcast.py

하위호환 (아무것도 깨지지 않음):
- 구 스크립트 경로는 deprecation 경고 후 새 모듈을 그대로 실행하는
  shim으로 유지 — 운영서버/구독자 crontab 무변경 동작
- loop_publish는 re-export shim 유지
- DB 테이블명(loop_a_position_state 등)은 상태 연속성 위해 불변
- env legacy 별칭(LOOP_A_*/LOOP_B_*/LOOP_C_*) fallback 유지
- feature_status는 신·구 파일명 모두 cron 매칭

로거명/로그 프리픽스([LOOP_C]->[FILL_CHASER])/내부 상수/테스트/문서
(FEATURE_FLAGS 용어집, OPS_HARNESS) 일괄 갱신.

테스트: 리네임 기인 신규 실패 0 (기존 main 실패 4건 중 1건은
FakeAgent **kwargs 수용으로 회복, 나머지 3+1건은 exit_kind/브로드캐스트
변경에 따른 기존 stale 테스트 — 본 PR 범위 외).
